Correct test timings; they were being recorded by the wrong test.  Closes RT#53927.
Previously, test timings were recorded "off by one" w.r.t. the actual
test in the JUnit output.  This commit corrects this and includes a set
of unit tests to ensure that timings are calculated correctly.

Thanks to Marc Abramowitz who submitted the original patch on which this
is based.