Capture and track timings for "initialization".
If your application takes a while to load, or performs compile-time
operations, this time was previously hidden; although we were tracking
time from one test to the next, we weren't capturing the load/compile
time for the test.

Without this, Hudson silently hides the time spent during init, as it
calculates the total time taken for a given test suite from the
component tests and not from the <testsuite> "time" attribute.