Capture and track time for test teardown.
Similarly to init/compile time tracking, teardown tracking captures
information on how long it took for things to shut down once we received
the last "test" in the TAP output.

If your test needed to do any kind of processing to clean up after
itself when the testing was finished, Hudson was not reporting that time
(as it calculates total test time by adding up each of the individual
tests, not by looking at the "<testsuite time=..>" attribute.