Skip bailout test on broken versions of `Test::Harness`
A bug was introduced in `Test::Harness` v3.45_01 which would cause
bail-out to produce a double summary in the output, which then broke our
tests.  This was then fixed in `Test::Harness` v3.50.

If we are run using a known broken version of `Test::Harness`, skip our
bail-out test.

Fixes #15